+++ app/Console/Commands/ConvertVideoCommand.php
@@ -8,7 +8,10 @@ use FFMpeg\Format\Video\WebM;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+<<<<<<< HEAD
 use Modules\Media\Support\Ffmpeg\MediaExporterResolver;
+=======
+>>>>>>> be7d0c3 (.)
 use ProtoneMedia\LaravelFFMpeg\Support\FFMpeg;
 use Webmozart\Assert\Assert;
 
@@ -31,7 +34,11 @@ class ConvertVideoCommand extends Command
             return '';
         }
 
+<<<<<<< HEAD
         $format = new WebM;
+=======
+        $format = new WebM();
+>>>>>>> be7d0c3 (.)
         $extension = mb_strtolower(class_basename($format));
         $file_new = Str::of($file)->replaceLast('.mp4', '.'.$extension)->toString();
 
@@ -43,10 +50,16 @@ class ConvertVideoCommand extends Command
             $this->info("{$remaining} seconds left at rate: {$rate}");
         });
 
+<<<<<<< HEAD
         $formattedExport = MediaExporterResolver::from(
             $export->toDisk($disk)
         )->inFormat($format);
         $formattedExport->save($file_new);
+=======
+        $export->toDisk($disk);
+        $export->inFormat($format);
+        $export->save($file_new);
+>>>>>>> be7d0c3 (.)
 
         return Storage::disk($disk)->url($file_new);
     }