+++ docs/changelog.md
@@ -1,5 +1,23 @@
+<<<<<<< HEAD
+=======
+---
+title: "Changelog"
+module: "Media"
+type: concept
+tags: [changelog]
+created: 2026-07-14
+updated: 2026-07-14
+qmd: "changelog"
+related:
+  - "./webm.md"
+---
+>>>>>>> be7d0c3 (.)
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
+<<<<<<< HEAD
 ## [Unreleased]
+=======
+## [Unreleased]
+>>>>>>> be7d0c3 (.)